:sparkles: Updated the About window
I added a row to show the space of the selected tool/launcher. I modified the row named Directory to Path since in my opinion it was better it was a better word for it. I updated the translation files since new text was added. Note that the new GetDirectorySize function is using a process to get the size with the 'du -sh $path' command. It's something I never used and I don't know how well it will work with Flatpak.