:lipstick: Modernised the UI
I modernized the user interface using the latest Adwaita widgets. I also updated the runtime version of the application to GNOME 45 to be able to use the latest Gtk/Adwaita widgets. I also  reorganized the files. I also made so that the desktop file and the appdata file are now translatable.