Fix download progress to prefer server's size and fix division by zero
The server always knows the true filesize, so we use it as the authority for the download progress whenever the server returns a value.

Also fixed a division by zero in the progress bar, which happened for things like GitHub's source downloads (they returned size 0 since the archive is created live on the server with an unknown size), which mean that it was doing "bytes_downloaded / 0" and causing a huge overflowing "23943458345%" percentage.

The progress will now always say "50%" when downloading an unknown size. This could later be enhanced to show actual download size (see source comment).